pyproject.toml
[tool.poetry]
name = "mlops_3"
version = "0.1.0"
description = "Training on MLOps"
authors = ["yeh-eng <you@example.com>"]
readme = "README.md"
license = "MIT"
packages = [{ include = "mlops_3", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
mlflow = "^2.20.3"
numpy = "^1.25"
pandas = "^2.1"
scikit-learn = "^1.3"
matplotlib = "^3.8"
ipykernel = "^6.29"
jupyter = "^1.0"
hydra-core = "^1.3.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-cov = "^6.0"
pytest-mock = "^3.14"
pytest-xdist = "^3.6"
black = "^23.9"
ruff = "^0.9.9"
mypy = "^1.15"
coverage = "^7.6"
bandit = "^1.8"
pre-commit = "^4.1"
commitizen = "^4.4"
pdoc = "^15.0"

[tool.poetry.scripts]
mlops_3 = "mlops_3.scripts:main"

[tool.commitizen]
name = "cz_conventional_commits"
version_scheme = "pep440"
tag_format = "v$version"
version_provider = "poetry"
update_changelog_on_bump = true

[tool.coverage.run]
branch = true
source = ["src"]
omit = ["__main__.py"]

[tool.mypy]
pretty = true
python_version = "3.10"
check_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "--verbosity=2"
pythonpath = ["src"]

[tool.ruff]
fix = true
indent-width = 4
line-length = 100
target-version = "py310"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["D100", "D103"]

[build-system]
requires = ["poetry-core>=1.5.1"]
build-backend = "poetry.core.masonry.api"
